+++ RepairShop/Areas/Admin/Pages/SerialNumbers/Upsert.cshtml.cs
@@ -195,7 +195,7 @@ namespace RepairShop.Areas.Admin.Pages.SerialNumbers
             };
         }
 
-        private async Task PopulateMaintenanceContracts(int clientId)
+        private async Task PopulateMaintenanceContracts(long clientId)
         {
             var contracts = (await _unitOfWork.MaintenanceContract.GetAllAsy(
                 mc => mc.IsActive == true && mc.ClientId == clientId,
